+++ tests/unit/sync/test_driver.py
@@ -38,8 +38,6 @@ from neo4j import (
     PreviewWarning,
     Query,
     Result,
-    TRUST_ALL_CERTIFICATES,
-    TRUST_SYSTEM_CA_SIGNED_CERTIFICATES,
     TrustAll,
     TrustCustomCAs,
     TrustSystemCAs,
@@ -144,17 +142,6 @@ def test_routing_driver_constructor(
         ({"encrypted": False}, ConfigurationError, '"encrypted"'),
         ({"encrypted": True}, ConfigurationError, '"encrypted"'),
         (
-            {"encrypted": True, "trust": TRUST_ALL_CERTIFICATES},
-            ConfigurationError,
-            '"encrypted"',
-        ),
-        ({"trust": TRUST_ALL_CERTIFICATES}, ConfigurationError, '"trust"'),
-        (
-            {"trust": TRUST_SYSTEM_CA_SIGNED_CERTIFICATES},
-            ConfigurationError,
-            '"trust"',
-        ),
-        (
             {"encrypted": True, "trusted_certificates": TrustAll()},
             ConfigurationError,
             '"encrypted"',
@@ -186,20 +173,13 @@ def test_routing_driver_constructor(
 def test_driver_config_error_uri_conflict(
     test_uri, test_config, expected_failure, expected_failure_message
 ):
-    def driver_builder(expect_failure=False):
-        if "trust" in test_config and not expect_failure:
-            with pytest.warns(DeprecationWarning, match="trust"):
-                return GraphDatabase.driver(test_uri, **test_config)
-        else:
-            return GraphDatabase.driver(test_uri, **test_config)
-
     if "+" in test_uri:
         # `+s` and `+ssc` are shorthand syntax for not having to configure the
         # encryption behavior of the driver. Specifying both is invalid.
         with pytest.raises(expected_failure, match=expected_failure_message):
-            driver_builder(expect_failure=True)
+            GraphDatabase.driver(test_uri, **test_config)
     else:
-        driver = driver_builder()
+        driver = GraphDatabase.driver(test_uri, **test_config)
         driver.close()
 
 
@@ -214,21 +194,6 @@ def test_driver_config_error_uri_conflict(
 def test_invalid_protocol(test_uri):
     with pytest.raises(ConfigurationError, match="scheme"):
         GraphDatabase.driver(test_uri)
-
-
-@pytest.mark.parametrize(
-    ("test_config", "expected_failure", "expected_failure_message"),
-    (
-        ({"trust": 1}, ConfigurationError, "The config setting `trust`"),
-        ({"trust": True}, ConfigurationError, "The config setting `trust`"),
-        ({"trust": None}, ConfigurationError, "The config setting `trust`"),
-    ),
-)
-def test_driver_trust_config_error(
-    test_config, expected_failure, expected_failure_message
-):
-    with pytest.raises(expected_failure, match=expected_failure_message):
-        GraphDatabase.driver("bolt://127.0.0.1:9001", **test_config)
 
 
 @pytest.mark.parametrize(